almost finished findtwins
Just need to finsh up teh ascii art

+++ SP1Framework/Scream.h
@@ -12,7 +12,7 @@ private:
 	//variables
 	bool gameends;
 	bool paused;
-	double power;
+	int power;
 	int currFrame;
 	bool lockinput;
 	Graphics& console;
